requirements.txt
openai>=1.0.0
prompt-toolkit>=3.0.0
pyyaml>=6.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
sqlite-vec==0.1.6